Update Jabbas Palace Fixes #50
