+++ app/(jabbalayout)/jabbaspalace/bib-fortuna/page.js
@@ -1,19 +1,14 @@
-import Layout, { siteTitle } from "./layout";
 import Image from "next/image";
-import jabbastyles from "./jabbastyles.module.css";
-import BibBoushh from "../../public/images/Jabba_Bib_Fortuna_Boushh.jpg";
-import BibFortunaPalace from "../../public/images/Jabba_Bib_Fortuna.jpg";
-import BibFortunaFace from "../../public/images/Jabba_BibFortuna_Face.png";
-import Head from "next/head";
+import BibBoushh from "../../../../public/images/Jabba_Bib_Fortuna_Boushh.jpg";
+import BibFortunaPalace from "../../../../public/images/Jabba_Bib_Fortuna.jpg";
+import BibFortunaFace from "../../../../public/images/Jabba_BibFortuna_Face.png";
+import jabbastyles from "../jabbastyles.module.css";
 
 export default function BibFortuna() {
-  const title = `${siteTitle} : Bib Fortuna`;
+  const title = ` : Bib Fortuna`;
 
   return (
-    <Layout>
-      <Head>
-        <title>{title}</title>
-      </Head>
+    <>
       <h1>Bib Fortuna</h1>
       <dl>
         <Image
@@ -98,6 +93,6 @@ export default function BibFortuna() {
         </dd>
         <dd>When this happened Bib Fortuna's plan was destined to fail.</dd>
       </dl>
-    </Layout>
+    </>
   );
 }
