+++ app/(totbh)/layout.js
@@ -1,9 +1,9 @@
 import Image from "next/image";
-import styles from "../components/layout.module.css";
-import utilStyles from "../styles/utils.module.css";
+import styles from "../../components/layout.module.css";
+import utilStyles from "../../styles/utils.module.css";
 import Link from "next/link";
-import "../styles/global.css";
-import MainNav from "./components/MainNav";
+import "../../styles/global.css";
+import MainNav from "../components/MainNav";
 
 const name = "Tales of the Bounty Hunters";
 
@@ -12,7 +12,7 @@ export const metadata = {
   description: "Blog for Paddy O'Hara",
 };
 
-export default function Layout({ children }) {
+export default function RootLayout({ children }) {
   return (
     <>
       <html>
